moved get_groups into the common_model because of the needed access in both home and group_model
