+++ app/hindsite/common_model.py
@@ -22,6 +22,20 @@ def get_user(email: str):
     return None
 
 
+def get_groups(email: str):
+    """
+    Gets all group records belonging to a user.
+
+    :param email: **str** Email to check against the database
+    :returns: **list** A list of the user's current groups.
+    """
+    user = get_user(email)
+    groups = []
+    for membership in user.groups:
+        if membership.invitation_accepted is True:
+            groups.append(membership.group)
+    return groups
+
 def get_group(group_id: int):
     """
     Gets a group record belonging to a user.
