+++ app/hindsite/group/templates/partials/search-results.html
@@ -10,7 +10,7 @@
                             {{ user.display_name }}
                         </p> 
                         <a href="#" class="px-5 py-1 btn btn-light">
-                            Invite User
+                            Send Invite
                             <img
                             class="img-fluid"
                             height="25"
